pyproject.toml
[project]
name = "teccod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "opensearch-py==2.7.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv==1.0.1",
    "uvicorn>=0.35.0",
]
